Add trip point presenter
